refactor config service
